r297: removed ce_softmax and ce_sigm
ce_softmax(y1,y0) is equivalent to ce_multi(softmax(x1),y0). The initial
incentive to have ce_softmax(y1,y0) is that its gradient is trivial to
calculate -- it's y1-y0. However, as we often need to compute the value, we
have to compute softmax() anyway inside ce_softmax(). This defeats the goal of
ce_softmax() and complicates the code.