Adds SIGINT catching on training loop
With this commit, during training, a user can ctr-C to send SIGINT. The signal
is caught and the looping variable is set to it's last valid value. This means
that the training loop finishes it's current instance and exits normally.

With this commit, users can decide to interrupt the training process at any
moment and till conserve the state of the model reached so far.